Fix `negative_sampling` when no negative samples are requested
`sample()` builds its result with `torch.tensor(random.sample(...))`, which
falls back to the default float dtype when the sampled list is empty. The
other branch returns `torch.arange(...)`, so the two branches disagree on
dtype as soon as `k == 0`.

Whenever the number of requested negative samples rounds down to zero, the
resulting index tensor is a float tensor. With `method="sparse"` that float
tensor propagates into the returned edge index, which is silently no longer
an integer tensor; with `method="dense"` it is used to index a mask and
raises `IndexError`.

This is reachable via `num_neg_samples=0`, via a float ratio that rounds
down to zero, and via `force_undirected=True` on a single edge, since that
halves the requested count.

Pass `dtype=torch.long` explicitly so both branches agree.

+++ torch_geometric/utils/_negative_sampling.py
@@ -326,7 +326,8 @@ def sample(
     if population <= k:
         return torch.arange(population, device=device)
     else:
-        return torch.tensor(random.sample(range(population), k), device=device)
+        return torch.tensor(random.sample(range(population), k),
+                            dtype=torch.long, device=device)
 
 
 def edge_index_to_vector(
